Revert "revert restructuring"
