+++ scripts/.gitignore
