+++ scripts/gpio_check.py
@@ -2,10 +2,9 @@
 # -*- coding: utf-8 -*-
 # author: Thomas Kaulke, kaulketh@gmail.com
 
-from __future__ import absolute_import
 from subprocess import Popen, PIPE, STDOUT, call
 
-import conf.greenhouse_config as config
+import greenhouse_config as config
 
 import os
 import sys
