+++ scripts/greenhouse_telegrambot.py
@@ -3,14 +3,12 @@
 # original: author: Stefan Weigert  http://www.stefan-weigert.de/php_loader/raspi.php
 # adapted: author: Thomas Kaulke, kaulketh@gmail.com
 
-from __future__ import absolute_import
-
 import logging
 import os
 import time
 
 import dht as dht
-import conf.greenhouse_config as conf
+import greenhouse_config as conf
 from telegram import (ReplyKeyboardMarkup,
                       ReplyKeyboardRemove, ParseMode)
 from telegram.ext import (Updater, CommandHandler, RegexHandler, ConversationHandler)
